a fix of line-ending code.
Before this fix: On Windows, if doing

  csv-find-dup.py ... -vvv > foo.txt

this will cause foo.txt's line separator to be '\r\r\n', which is not desired. This is due to -vvv implies binary-mode reading of the input csv.

Now, the line separator will be '\r\n'.

CAUSE: print("abc\r\n", end='') will result in '\r\r\n' in foo.txt .